Import programs
Users import a program into the wallet to use it securely in the asm
fragment. The wallet will reject asm fragments with unknown CMRs.
Under the hood, the wallet just records the CMR of the main program.
This is because disconnected expressions might not be available at
commitment time. The user is expected to hold onto the program and fill
in the disconnected expressions (holes) at commitment time.