Assembly: Store satisfactions
RedeemNode does not have a serde implementation in rust-simplicity
because the base64 dependency is missing. As a fix, I create a wrapper
that implements serde in the wallet. This is also useful because we want
to (de)serialize WitnessNode, which might never get a serde
implementation in rust-simplicity. In our case, we know that all witness
data and disconnected branches are populated, so it is safe to
(de)serialize. In the future, we can think about writing our own base64
encoding logic in rust-simplicity.