UTXO: Store descriptor
Storing the UTXO descriptor is more general than storing the index of
the child key. This lets us handle assembly fragments and other use
cases.
As a side effect, the signing logic gets simpler.
We change the assembly descriptors to PublicKey for compatability with
the xpub descriptors. For asm the key type doesn't matter. xpub children
can be converted into PublicKey.